requirements.txt

burnysc2
numpy
loguru